refactoring, and general clean up
